Converted Books to Use New Sqlite Async Library
